fix: fix some furnace bugs

+++ Allay-Server/src/main/java/org/allaymc/server/blockentity/component/furnace/BlockEntityBlastFurnaceBaseComponentImpl.java
@@ -31,7 +31,7 @@ public class BlockEntityBlastFurnaceBaseComponentImpl extends BlockEntityFurnace
     }
 
     @Override
-    public int getSpeedWhenFurnaceTypeMostSuitable() {
+    public float getSpeedWhenFurnaceTypeMostSuitable() {
         return 2;
     }
 }
